requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
plotly>=5.18.0
fpdf2>=2.7.0
azure-identity>=1.15.0
azure-mgmt-authorization>=4.0.0
